Split up BeanMapperTest into BeanMapperMockTest
the tests use different harnesses and there's multiple shadowed variables
