+++ core/src/test/java/org/jdbi/v3/core/mapper/reflect/BeanMapperTest.java
@@ -14,258 +14,35 @@
 
 package org.jdbi.v3.core.mapper.reflect;
 
-import java.math.BigDecimal;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
 import javax.annotation.Nullable;
 
 import org.jdbi.v3.core.Handle;
-import org.jdbi.v3.core.HandleAccess;
 import org.jdbi.v3.core.SampleBean;
 import org.jdbi.v3.core.Something;
-import org.jdbi.v3.core.ValueType;
 import org.jdbi.v3.core.mapper.Nested;
 import org.jdbi.v3.core.mapper.PropagateNull;
 import org.jdbi.v3.core.mapper.RowMapper;
-import org.jdbi.v3.core.mapper.ValueTypeMapper;
 import org.jdbi.v3.core.rule.H2DatabaseRule;
-import org.jdbi.v3.core.statement.StatementContext;
-import org.jdbi.v3.core.statement.StatementContextAccess;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.when;
 
 public class BeanMapperTest {
-
-    @Rule
-    public MockitoRule mockitoRule = MockitoJUnit.rule();
-
     @Rule
     public H2DatabaseRule dbRule = new H2DatabaseRule().withSomething();
 
-    @Mock
-    ResultSet resultSet;
-
-    @Mock
-    ResultSetMetaData resultSetMetaData;
-
-    Handle handle = HandleAccess.createHandle();
-    StatementContext ctx = StatementContextAccess.createContext(handle);
+    Handle handle;
 
     RowMapper<SampleBean> mapper = BeanMapper.of(SampleBean.class);
 
     @Before
-    public void setUpMocks() throws SQLException {
-        when(resultSet.getMetaData()).thenReturn(resultSetMetaData);
-    }
-
-    private void mockColumns(String... columns) throws SQLException {
-        when(resultSetMetaData.getColumnCount()).thenReturn(columns.length);
-        for (int i = 0; i < columns.length; i++) {
-            when(resultSetMetaData.getColumnLabel(i + 1)).thenReturn(columns[i]);
-        }
-    }
-
-    private void mockLongResult(long aLong) throws SQLException {
-        when(resultSet.getLong(1)).thenReturn(aLong);
-        when(resultSet.wasNull()).thenReturn(false);
-    }
-
-    private void mockAllNullsResult() throws SQLException {
-        when(resultSet.wasNull()).thenReturn(true);
-    }
-
-    @Test
-    public void shouldSetValueOnPublicSetter() throws Exception {
-        mockColumns("longField");
-
-        Long aLongVal = 100L;
-        mockLongResult(aLongVal);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-    }
-
-    @Test
-    public void shouldHandleColumNameWithUnderscores() throws Exception {
-        mockColumns("LONG_FIELD");
-
-        Long aLongVal = 100L;
-        mockLongResult(aLongVal);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-    }
-
-    @Test
-    public void shouldBeCaseInSensitiveOfColumnWithUnderscoresAndPropertyNames() throws Exception {
-        mockColumns("LoNg_FiElD");
-
-        Long aLongVal = 100L;
-        mockLongResult(aLongVal);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-    }
-
-    @Test
-    public void shouldHandleEmptyResult() throws Exception {
-        mockColumns();
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean).isNotNull();
-    }
-
-    @Test
-    public void shouldBeCaseInSensitiveOfColumnAndPropertyNames() throws Exception {
-        mockColumns("LoNgfielD");
-
-        Long aLongVal = 100L;
-        mockLongResult(aLongVal);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(aLongVal);
-    }
-
-    @Test
-    public void shouldHandleNullValue() throws Exception {
-        mockColumns("LoNgfielD");
-
-        mockAllNullsResult();
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isNull();
-    }
-
-    @Test
-    public void shouldSetValuesOnPublicSetter() throws Exception {
-        mockColumns("longField");
-
-        Long expected = 1L;
-        mockLongResult(expected);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(expected);
-    }
-
-    @Test
-    public void shouldThrowOnTotalMismatch() throws Exception {
-        mockColumns("somethingElseEntirely");
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldThrowOnProtectedSetter() throws Exception {
-        mockColumns("protectedStringField");
-
-        String expected = "string";
-        when(resultSet.getString(1)).thenReturn(expected);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldThrowOnPackagePrivateSetter() throws Exception {
-        mockColumns("packagePrivateIntField");
-
-        when(resultSet.getInt(1)).thenReturn(200);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldThrowOnPrivateSetter() throws Exception {
-        mockColumns("privateBigDecimalField");
-
-        when(resultSet.getBigDecimal(1)).thenReturn(BigDecimal.ONE);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldSetValuesInSuperClassProperties() throws Exception {
-        mockColumns("longField", "blongField");
-
-        Long aLongVal = 100L;
-        Long bLongVal = 200L;
-
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.getLong(2)).thenReturn(bLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        RowMapper<DerivedBean> mapper = BeanMapper.of(DerivedBean.class);
-
-        DerivedBean derivedBean = mapper.map(resultSet, ctx);
-
-        assertThat(derivedBean.getLongField()).isEqualTo(aLongVal);
-        assertThat(derivedBean.getBlongField()).isEqualTo(bLongVal);
-    }
-
-    @Test
-    public void shouldUseRegisteredMapperForUnknownPropertyType() throws Exception {
-        handle.registerColumnMapper(new ValueTypeMapper());
-
-        mockColumns("longField", "valueTypeField");
-        Long expected = 123L;
-
-        when(resultSet.getLong(1)).thenReturn(expected);
-        when(resultSet.getString(2)).thenReturn("foo");
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(expected);
-        assertThat(sampleBean.getValueTypeField()).isEqualTo(ValueType.valueOf("foo"));
-    }
-
-    @Test
-    public void shouldThrowOnPropertyTypeWithoutRegisteredMapper() throws Exception {
-        mockColumns("longField", "valueTypeField");
-
-        when(resultSet.getLong(1)).thenReturn(123L);
-        when(resultSet.getObject(2)).thenReturn(new Object());
-        when(resultSet.wasNull()).thenReturn(false);
-
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void shouldNotThrowOnMismatchedColumns() throws Exception {
-        mockColumns("longField", "extraColumn");
-
-        Long expected = 666L;
-        when(resultSet.getLong(1)).thenReturn(expected);
-        when(resultSet.getString(2)).thenReturn("foo");
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertThat(sampleBean.getLongField()).isEqualTo(expected);
-    }
-
-    @Test
-    public void shouldThrowOnMismatchedColumnsStrictMatch() throws Exception {
-        ctx.getConfig(ReflectionMappers.class).setStrictMatching(true);
-        mockColumns("longField", "misspelledField");
-        assertThatThrownBy(() -> mapper.map(resultSet, ctx)).isInstanceOf(IllegalArgumentException.class);
+    public void getHandle() throws SQLException {
+        handle = dbRule.getSharedHandle();
     }
 
     public static class ColumnNameBean {
@@ -293,7 +70,6 @@ public class BeanMapperTest {
 
     @Test
     public void testColumnNameAnnotation() {
-        Handle handle = dbRule.getSharedHandle();
         handle.registerRowMapper(BeanMapper.factory(ColumnNameBean.class));
 
         handle.execute("insert into something (id, name) values (1, 'foo')");
@@ -308,7 +84,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNested() {
-        Handle handle = dbRule.getSharedHandle();
         handle.registerRowMapper(BeanMapper.factory(NestedBean.class));
 
         handle.execute("insert into something (id, name) values (1, 'foo')");
@@ -323,7 +98,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNestedStrict() {
-        Handle handle = dbRule.getSharedHandle();
         handle.getConfig(ReflectionMappers.class).setStrictMatching(true);
         handle.registerRowMapper(BeanMapper.factory(NestedBean.class));
 
@@ -346,7 +120,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNestedNotReturned() {
-        Handle handle = dbRule.getSharedHandle();
         handle.registerRowMapper(BeanMapper.factory(NestedBean.class));
         assertThat(handle
             .createQuery("select 42 as testValue")
@@ -380,7 +153,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNestedPrefix() {
-        Handle handle = dbRule.getSharedHandle();
         handle.registerRowMapper(BeanMapper.factory(NestedPrefixBean.class));
 
         handle.execute("insert into something (id, name) values (1, 'foo')");
@@ -395,7 +167,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNestedPrefixStrict() {
-        Handle handle = dbRule.getSharedHandle();
         handle.getConfig(ReflectionMappers.class).setStrictMatching(true);
         handle.registerRowMapper(BeanMapper.factory(NestedPrefixBean.class));
 
@@ -425,7 +196,6 @@ public class BeanMapperTest {
 
     @Test
     public void testNestedPrefixNotReturned() {
-        Handle handle = dbRule.getSharedHandle();
         handle.registerRowMapper(BeanMapper.factory(NestedPrefixBean.class));
         assertThat(handle
             .createQuery("select 42 as integerValue")
@@ -458,8 +228,7 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void testPropagateNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void propagateNull() {
         assertThat(handle
             .registerRowMapper(BeanMapper.factory(PropagateNullThing.class))
             .select("SELECT null as testValue, 'foo' as s")
@@ -469,8 +238,7 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void testPropagateNotNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void propagateNotNull() {
         assertThat(handle
             .registerRowMapper(BeanMapper.factory(PropagateNullThing.class))
             .select("SELECT 42 as testValue, 'foo' as s")
@@ -481,8 +249,7 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void testNestedPropagateNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void nestedPropagateNull() {
         assertThat(handle
             .registerRowMapper(BeanMapper.factory(NestedPropagateNullThing.class))
             .select("SELECT 42 as integerValue, null as testValue, 'foo' as s")
@@ -493,8 +260,7 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void testNestedPropagateNotNull() {
-        Handle handle = dbRule.getSharedHandle();
+    public void nestedPropagateNotNull() {
         assertThat(handle
             .registerRowMapper(BeanMapper.factory(NestedPropagateNullThing.class))
             .select("SELECT 42 as integerValue, 60 as testValue, 'foo' as s")
